Correct editing bug when editing as group owner
When creating a new newsletter as a non-admin user (ie. group admin or group operator), creation fails with "IOException:UnableToSaveNew" exception.

I haven't checked deeply enough to tell why this doesn't happen when creating it with an admin user, but anyway this small patch corrects it by setting the access_id before saving (which is required by ElggEntity->save() function).